Make syntect global only
