+++ rust/core-lib/assets/client_example.toml
@@ -26,7 +26,6 @@ font_face = "InconsolataGo"
 font_size = 14
 
 # Automatically match current indentation level on newline.
-# note: this is currently broken on files with > 1k lines
 auto_indent = false
 
 # Allow scrolling past the last line of a document.
