+++ rust/syntect-plugin/Cargo.toml
@@ -9,10 +9,6 @@ description = "A syntax highlighting plugin based on syntect."
 [dependencies]
 serde = "1.0"
 serde_derive = "1.0"
-serde_json = "1.0"
-
-[features]
-global = []
 
 [dependencies.syntect]
 version = "2.0"
